fix bugs :construction:
